requirements.txt
Django==3.1.7
PyMySQL==1.0.2
djangorestframework==3.11.0
mysqlclient==1.4.6
